localdatetime를 받아오지 않고, utc로 받아오던 현상 수정완료

+++ vehicle-tracker-backend/vehicle-tracker-backend/src/main/kotlin/com/example/vehicle_tracker_backend/Model/VehicleStatus.kt
@@ -8,7 +8,7 @@ data class VehicleStatus(
     val deviceId: String,
     val bluetoothDevice: String,
     val engineStatus: String,
-    val speed : Double,
+    val speed : Int,
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     val timestamp: LocalDateTime = LocalDateTime.now(),
     val location: VehicleLocation? = null
